Split query history file
