+++ extensions/ql-vscode/src/remote-queries/variant-analysis-content-provider.ts
@@ -1,7 +1,7 @@
 import { TextDocumentContentProvider, Uri } from "vscode";
 import { URLSearchParams } from "url";
 import { showAndLogWarningMessage } from "../helpers";
-import { SHOW_QUERY_TEXT_MSG } from "../query-history/query-history";
+import { SHOW_QUERY_TEXT_MSG } from "../query-history/query-history-manager";
 import { VariantAnalysisManager } from "./variant-analysis-manager";
 
 export const createVariantAnalysisContentProvider = (
