+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-history/remote-query-history.test.ts
@@ -20,7 +20,6 @@ import {
 import { QueryHistoryConfig } from "../../../../src/config";
 import { DatabaseManager } from "../../../../src/databases";
 import { tmpDir, walkDirectory } from "../../../../src/helpers";
-import { QueryHistoryManager } from "../../../../src/query-history/query-history";
 import { AnalysesResultsManager } from "../../../../src/remote-queries/analyses-results-manager";
 import { RemoteQueryResult } from "../../../../src/remote-queries/shared/remote-query-result";
 import { DisposableBucket } from "../../disposable-bucket";
@@ -34,6 +33,7 @@ import { VariantAnalysisManager } from "../../../../src/remote-queries/variant-a
 import { App } from "../../../../src/common/app";
 import { createMockApp } from "../../../__mocks__/appMock";
 import { testCredentialsWithStub } from "../../../factories/authentication";
+import { QueryHistoryManager } from "../../../../src/query-history/query-history-manager";
 
 // set a higher timeout since recursive delete may take a while, expecially on Windows.
 jest.setTimeout(120000);
